fix: ディレクトリ構成を整理
