+++ iOSEngineerCodeCheck/Data/Infrastructure/RestApi/HTTPMethod.swift
@@ -1,0 +1,16 @@
+//
+//  HTTPMethod.swift
+//  iOSEngineerCodeCheck
+//
+//  Created by Shusuke Ota on 2025/1/26.
+//  Copyright © 2025 YUMEMI Inc. All rights reserved.
+//
+
+import Foundation
+
+enum HTTPMethod: String {
+    case get    = "GET"
+    case post   = "POST"
+    case put    = "PUT"
+    case delete = "DELETE"
+}
